contributors/i18n/submitted/css3-syntax/i18n-css-syntax-bytes-008.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Syntax, input byte stream: no semicolon at end of charset rule</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-syntax/#input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http dom'>
<meta name="assert" content="If a @charset declaration is missing a final semicolon, the encoding declaration will not be recognised.">
<style type='text/css'>
.test div { width:50px; }

p { color: white; margin-top: 3em; font-size: 80%; }
</style>
<link rel="stylesheet" type="text/css" href="support/charset15-nocolon.css">
</head>
<body>



<div class="test"><div id="box" class="ÃœÃ€Ãš">&#xA0;</div></div>


<!--Notes:
This test is only valid if the user agent recognises the @charset rule.

In the style sheet, the CSS selector identifies a class name with the following byte sequence: c3 bd c3 a4 c3 a8.  Read as UTF-8 this gives the characters ýäè.  Read as ISO 8859-1, this gives  Ã½Ã¤Ã¨. Read as ISO 8850-15, this gives ÃœÃ€Ãš.

This HTML page is served as UTF-8 and uses a different byte sequence for the class name (c3 83 c5 93 c3 83 e2 82 ac c3 83 c5 a1) that equates to the character sequence ÃœÃ€Ãš.  

The class ids will therefore only match if the CSS is read as ISO 8859-15.

There is an @charset rule at the start of the style sheet that can set the encoding to ISO 8859-15, but it is incorrectly formed, and so should be ignored. If the previously mentioned selector matches, ie. the @charset rule has not been ignored, the test fails.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, " ");
</script>

<div id='log'></div>

<p class="ÃœÃ€Ãš"><span><span id="debug">Text as read from the style sheet: </span><br />(If blank, the style sheet hasn't been read at all.)</span></p>
</body>
</html>
